Refactor: Move experiment log code to separate module
